requirements.txt
bcrypt==3.1.7
beautifulsoup4==4.8.0
cffi==1.12.3
Click==7.0
django-htmlmin==0.11.0
Flask==1.1.1
Flask-Static-Compress==1.0.2
Flask-WTF==0.14.2
html5lib==1.0.1
itsdangerous==1.1.0
jac==0.17.2
Jinja2==2.10.1
MarkupSafe==1.1.1
ordereddict==1.1
passlib==1.7.1
pycparser==2.19
pycrypto==2.6.1
PyMySQL==0.9.3
rjsmin==1.1.0
simple-crypt==4.1.7
six==1.12.0
soupsieve==1.9.3
webencodings==0.5.1
Werkzeug==0.15.6
WTForms==2.2.1
